tb: add `impl std::error::Error for Error`
This commit adds an implementation of the `std::error::Error` trait for
the main `Error` class.

As `std::error` is not available in `core` yet the feature is gated
with `error_trait` keyword in the specification file. By default the
feature is enabled (as `no_std` is by default disabled in Rust).

To disable the feature in `no_std` environment, or in order to provide a
custom implementation of the trait, use `error_trait: false` in the
(Yaml) specification file.