tb/parser: add stricter name checking and disallow whitespace
This commit makes the error name check stricter, because previously a
name with spaces was not rejected as an invalid name.

From now on only [A-Za-z0-9_] characters are allowed in a name.

Additionally this commit extacts the generic check of an identifier into
a separate `check_ident` function. This is done in preparation for
customizable struct names which will also need a similar check.