tighterror: rename `ErrorCode` to `ErrorKind`
Initially I named it `ErrorCode`. However, when I started to write the
framework documentation I realized that making it more standard-library
alike is a better idea, cause the concepts will be more familiar to
users.

This commit renames ErrorCode to ErrorKind. What was previously named
error-kind, i.e., the ID of an error within a category, is now called
error-variant.

This commit is a large change, also updating the documentation and
variable names where appropriate. Additionally, this commit refactors
bit-shift operations to avoid undefined behavior.