tb: add `update` mode
In `update` mode the destination file isn't overwritten if the new
to-be-written data equals the existing one. This is handy in the
makefile to avoid changing the target module file's modification time,
which causes a crate recompilation. The default behavior remains
unchanged, i.e., the destination file is overwritten unconditionally.