ct: add `update` mode
This commit adds a new argument `-u, --update` which enables the update
mode. In update mode the destination file isn't overwritten if the
to-be-written data equals the existing one. This way modification of the
target file is avoided in many cases, which avoids an unnecessary
recompilation of a user crate.