makefile: add `up` target
The new `up` target generates the tighterror-build errors in `update`
mode. This is done to avoid updating the file's modification time if the
new data is the same as the existing one. This prevents unnecessary
rebuilds of the crate as cargo triggers a rebuild based on source files'
modification time.