.gitignore: ignore the `target` directory in all subdirectories
This is needed because some tests in TB are not part of the main
workspace. Therefore, when they are built a `target` directory is
created in their corresponding root folder. Till now `target` was
ignored only in the root directory of the repo. Now it will be ignored
in those subdirectories too.