tb: extract module attributes to a separate object under `module` keyword
This commit introduces a new Module Object in preparation for
supporting multiple tighterror modules in the same specification file.

Note that this is a breaking change. It places module-specific
attributes under a new type of object, the Module Object. This means
that the [tighterror] section from now on contains only global attributes,
those affecting the entire code generation process.
Module-specific attributes, e.g., `err_cat_name`, are placed under a new
[module] section.

Note that `mod_doc` attribute has been removed, and the plain `doc`
attribute is used instead. Moreover, the `tighterror` keyword used
previously for module configuration is renamed to `main` and the
corresponding object is called the Main Object.