tb: properly handle multiline comments
The issue with multiline comments is that the `quote` crate simply
places them inside /** */ comment which by markdown rules is considered
as a code block when indented by more than 4 spaces.

I have reported this in cargo project, and the maintainer has explained
root cause of this behavior [1]. As far as I understand from comments on
issue [2] this behavior is not going to change soon in rustdoc.

This commit works around the issue by adding leading " * " to every line
in a multiline doc string. This way `quote` takes the string as-as and
indented code block is avoided.

[1] https://github.com/rust-lang/rust/issues/123853
[2] https://github.com/rust-lang/rust/issues/100225