tb: add support for custom categories
Note that this commit is not backward compatible because kind constants
are now placed in a dedicated module per category. So a kind constant
with an implicit category that was previously named kinds::BAD_FILE will
now be named kinds::general::BAD_FILE. This is done because error names
are required to be unique within a category, but distinct categories
may have errors with the same name.