tb: add module-level `flat_kinds` attribute
The new attribute `flat_kinds` affects the layout of the `kinds`
submodule. By default the `kinds` submodule has an inner submodule per
category which contains the category's error kind constants. This is
done because error names aren't required to be unique per module, but
only per category. This however, creates a nested layout of the `kinds`
module.

To avoid per-category inner modules, and thus making error kind
constants' paths shorter, this commit introduces the `flat_kinds`
module-level attribute. When enabled the error kind constants are placed
directly under the `kinds` submodule (as was done before support for
multiple categories was introduced), and inner per-category submodules
are skipped. The limitation is that when `flat_kinds` is enabled error
names are required to be unique per module and not per category only.