tb: coding: refactor `ModuleGenerator::calc_repr_type`
The function is always called with `n_kind_bits` parameter, which is
already verified to be within bounds on line 49. Hence, the input to
this function can never exceed 64, making the error condition
unreachable. This commit make the function infallible to avoid the need
to return an error code. The function panics instead to make the
compiler happy.