tb: handle module doc string for single module
Commit 4e0c4c0 has added support for multiple modules. Initially I
decided to ignore the `doc` attribute of a module when there is only a
single module in specification. The reason was that in case of a single
module tighterror doesn't wrap it in `mod <module_name> {}` block.
So documentation of the module can be added by the user explicitly.

However, the new behavior is unexpected in case one has more than one
module initially and then removes all but one. In this case if a `doc`
attribute is present the expectation is to have the doc string in
generated code.

This commit restores the old behavior, so now module doc string is
always generated if present in a specification.

tighterror-build's specification file is restored to have a doc string
as was before commit 4e0c4c0.