tb: fix treatment of a relative `output` path
The `output` specification attribute was designed to be either an
absolute path, or a path relative to the directory containing the
specification file itself.

This commit implements the handling of the relative path. Till now if
the path was relative it was taken relative to the current working
directory. Now it is relative to the location of the specification file.