tighterror: allow `MainObject::output` to point to a directory
This commit changes the definition of `MainObject::output` attribute to
support directory paths. If `output` points to a directory the output is
written to file `tighterror.rs` under that directory.

This change is done in preparation for supporting *separate files*
mode. In this mode every module is written to a separate file named
after the module. The `output` in this case will be required to point
to an existing directory.