tb: put category bits closer to most-significant bits
This commit conceptually reverts commits c99fb3f and c708d54.

The purpose of this commit is to put category bits higher in
significance order. This was the initial implementation until in commit
c99fb3f I switched the order to put category bits in LSB.

Commit c99fb3f mentions 3 benefits to having the order switched.
This commit reconsiders some of those benefits and provides an
alternative implementation to cancel others:

1. "Simplified calculation of variant and category values"
   an alternative implementation is provided without using checked
   rotation operators.

2. "Reduced amount of constants"
   looking at the roadmap, e.g., supporting Flags, constants removed in
   commit c99fb3f will be defined again in future.

3. "Putting variant bits next to reserved ones creates a continuous
    range of spare bits"
   considering the Flags feature again I do not see how unused variant
   bits are used without severely complicating calculation of
   `variant_value()` function. I prefer to keep it simple.

Another benefit of putting category bits closer to MSB, as I see it
now, is that it creates order between categories such that when sorted
all errors of one category have higher values than all other errors
having lower category id. In current implementation sorting all errors
in a module doesn't yield separation by category. This shouldn't be an
issue on its own as tighterror users shouldn't rely on errors having
any specific numeric value. But this property is aligned with the
concept of "category" that groups the errors logically. Having the
errors grouped numerically too seems like the right thing to do.