tb: require error names in UPPER_SNAKE case instead of UpperCamel
The purpose of this task is to unify the way errors are specified in
markup language and in Rust code.

Till now error names were required to be specified in UpperCammel case.
This is more sutable for struct names in Rust and is less suitable for
constant names, which are by Rust convention written in UPPER_SNAKE
case.

The benefit of having the error names specified in UPPER_SNAKE case is
that it becomes more greppable, i.e., when an error is grepped for in a
codebase the results will include the usage of the error constant as
well as the error definition in markup language.

Looking at the roadmap, this will not always be true when VariantTypes
are introduced. However, every variant type will be associated with a
error-kind constant which will be returned in grep results.