tb: add support for Variant Types
Variant Types are ZSTs created for error kinds. They are useful when a
function has only a single error condition. In such cases returning a
full-fledged Error may be excessive because that forces a caller to
handle a non-exhaustive list of errors. With Variant Types the return
code is exhaustive, making error handling simpler.

Note that Variant Types are marked `#[non_exhaustive]` to forbid their
creation outside their origin crate. This doesn't affect a function that
returns such type. A caller of such function still needs to handle only
a single error.

Variant Types may be enhanced in future to contain custom fields. This
will allow returning rich information about errors.

Variant Types are implicitly convertible to Kind and Error so
propagating them through layers of abstraction is seamless as tighterror
aims to be.