tb: add a casing test-case
To explicitly show the conversion end result from UpperCamel to
UpperCamel when the string's actual case is UPPER_SNAKE.

Specifically, the underscore is left there as a normal character and
lowerUpper boundary doesn't kick-in to detect ERR as a separate word.
So, the whole result appears as a single word.