tb: disallow underscore in UpperCamel case identifiers
Previously `check_ident_chars` was unaware of the casing of the
identifier, allowing all of `[A-Za-z0-9_]` characters in all
identifiers. However, UpperCamel case identifiers, e.g. a struct name,
shouldn't be delimited by underscore. `convert_case` treats
underscore as a regular character when checking compliance with
UpperCamel case, so it is not rejected. Therefore, this commit rejects
underscore at character checking stage.