tb: use name-constant instead of repeating the name-string in display constant definition
This commit changes the generator to reuse the already defined named
constant instead of repeating the name string in the definition of
display string. This is done to try to use a single definition for
every error property. Moreover, this is done in preparation of using the
name-constant for Variant Error Type display-constant, so that both
display constants reuse the already existing name definition and not
repeat the same string twice.

Due to reuse of name-constants in other sub-modules the per-category
name-constants modules and constants themselves are made `pub(crate)`
(previously they were private). Note, however, the names module `_n` is
private so those definitions aren't exposed crate-wise.